Crate icaros

ICAROS (In-flight Communications and Reporting Operations System), drone side.

Modules

comms
Communication systems

core
Core routing facilities and base logic

esb
External Service Broker

hardware
Secondary hardware support (i.e. camera, GPS, serial port)

Constants

CONTROLLER_SERIAL_PORT_PATH
Path to the serial port file where the flight controller will be connected to. MODIFY AS REQUIRED.

GPS_SERIAL_PORT_PATH
Path to the serial port file where the GPS module will be connected to. MODIFY AS REQUIRED.

Functions

main
Entry point for the program