Module icaros::comms::air

Air-to-ground communication

Modules

dummy
Dummy implementation of an air-to-ground communication service (used to test system's architecture)

rf24
nRF24 implementation of an air-to-ground communication service (used in production)

simulated
Simulated implementation of an air-to-ground communication service (used to test data sending/reception to the flight controller without a proper ground control present)

udp
UDP implementation of an air-to-ground communication service (used to test communication with ground control without a nRF24L01+ radio set up.