Module icaros::comms::application

In-flight application communication

Structs

ApplicationClientMessage
A complete message sent through the TCP socket by an application to the ICAROS system

ApplicationConnection
The TCP connection and associated state data between an application and the ICAROS system

ApplicationID
Represents an unique ID that identifies a single application running on the drone and communicating with the ICAROS system

ApplicationService
The application service

RequestID
Represents an unique ID that identifies a single request received from a given application running on the drone and communicating with the ICAROS system

Enums

ApplicationClientMessageType
The payload type of a complete message sent through the TCP socket by an application to the ICAROS system

ApplicationServiceMessage
A complete message sent through the TCP socket by the ICAROS system itself to the application

DataRecvResult
The result of attempting to receive a certain amount of data of type T through the TCP socket

Constants

READ_LOOP_LIMIT
The amount of loops that will be done within a single iteration to attempt to read messages

_IMPL_DESERIALIZE_FOR_ApplicationClientMessage

_IMPL_DESERIALIZE_FOR_ApplicationClientMessageType

_IMPL_DESERIALIZE_FOR_ApplicationID

_IMPL_DESERIALIZE_FOR_RequestID

_IMPL_SERIALIZE_FOR_ApplicationClientMessage

_IMPL_SERIALIZE_FOR_ApplicationClientMessageType

_IMPL_SERIALIZE_FOR_ApplicationID

_IMPL_SERIALIZE_FOR_ApplicationServiceMessage

_IMPL_SERIALIZE_FOR_RequestID