Struct icaros::comms::controller::scheduler::ControllerPacketScheduler

pub struct ControllerPacketScheduler<T: ControllerCommunicationService> {
    comm_service: T,
    tx_in: Option<R2AMessage>,
    latency_recv: u32,
    counter: u16,
}

Packet scheduler for the flight controller

T defines the underlying device driver to which packets are scheduled

Fields
comm_service: Ttx_in: Option<R2AMessage>latency_recv: u32counter: u16
Methods

impl<T: ControllerCommunicationService> ControllerPacketScheduler<T>

pub fn initialize(
config: T::ControllerCommunicationOptions
) -> Result<Self, T::HardwareDriverError>

Initializes a new flight controller packet scheduler, together with its underlying driver and given the latter's settings to be used

fn generate_message_from_state(&mut self, state: &SystemState)

Generates a new outgoing message from the given system state and schedules them for sending

fn update_state_from_message(
&mut self,
state: &mut SystemState,
message: A2RMessage
)

Updates the given system state from the given incoming message