Module icaros::comms::controller::serial

Serial port implementation of a flight controller communication service (used in production)

Structs

SerialControllerCommunicationsService
Serial-port-based flight controller communication service

Enums

ParsingState
Current frame parsing state

SerialCommunicationError
An error that occurred on the serial communication

Constants

SERIAL_END_BYTE
Byte value used to signal the end of a data frame

SERIAL_ESCAPE_BYTE
Byte value used to escape the next byte value from being interpreted in a special way

SERIAL_START_BYTE
Byte value used to signal the start of a data frame