Struct icaros::core::ICAROSController

pub struct ICAROSController<A, B> where
    A: ControllerCommunicationService,
    B: AirCommunicationService<A2GMessage, G2AMessage>,  {
    sys_data: SystemState,
    controller_scheduler: ControllerPacketScheduler<A>,
    air_scheduler: AirPacketScheduler<B, SystemState, A2GMessage, G2AMessage>,
    gps_device: Option<GPSService>,
    esb: ExternalServiceBroker,
}

The central controller for all the systems around ICAROS. Takes care of updating each one of them individually and of handling the state that is to be passed around to each system.

Fields
sys_data: SystemState

The system state

controller_scheduler: ControllerPacketScheduler<A>

The packet scheduler for the controller side

air_scheduler: AirPacketScheduler<B, SystemState, A2GMessage, G2AMessage>

The packet scheduler for the ground control side

gps_device: Option<GPSService>

The GPS service (optional)

esb: ExternalServiceBroker

The external service broker

Methods

impl<A, B> ICAROSController<A, B> where
A: ControllerCommunicationService,
B: AirCommunicationService<A2GMessage, G2AMessage>,

pub fn init(
controller_sched: ControllerPacketScheduler<A>,
air_sched: AirPacketScheduler<B, SystemState, A2GMessage, G2AMessage>,
gps_device: Option<GPSService>
) -> Result<Self, Box<dyn Error>>

Initializes the central controller with the given ground control and flight controller packet schedulers and, optionally, an initialized GPS device driver. May fail if the external service broker fails to start.

pub fn process(&mut self)

Processes a single tick. This will handle, in order:

Receiving new state from the flight controller

Receiving new state from the ground control

Route incoming application-specific packets from the ground control to the external service broker

If applicable, receiving current positional data from the GPS service

Send new state to the flight controller

Send new state to the ground control

Route outgoing application-specific messages to the ground control to the external service broker

pub fn get_stats(&self) -> &SystemStatistics

Gets the system's current statistics

pub fn reset_stats(&mut self)

Resets the system's current statistics